requirements.txt
requests
kafka-python
psycopg2-binary
pandas
numpy
scikit-learn
matplotlib